Revert "Modifications taken from main branch"
This reverts commit 485aab28a2019d9aa3157972655d5382bee38692.

+++ Application/EdFi.Admin.DataAccess/Utils/IDefaultApplicationCreator.cs
@@ -12,6 +12,6 @@ namespace EdFi.Admin.DataAccess.Utils
     {
         Application FindOrCreateUpdatedDefaultSandboxApplication(int vendorId, SandboxType sandboxType);
 
-        void AddEdOrgIdsToApplication(IList<long> edOrgIds, int applicationId);
+        void AddEdOrgIdsToApplication(IList<int> edOrgIds, int applicationId);
     }
 }
